feat: ui manager
