+++ Assets/Prefabs/UI/InventoryPanel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: da55978cfd3faf041ad98a666f4856da
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
