+++ Assets/ScriptableObjects/EventsChannel/Inventory/OnInventoryClose.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1c11e13bb4944b7468b781bd6695c380
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
