+++ Assets/ScriptableObjects/EventsChannel/Inventory/OnInventoryOpen.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: cbde39bc75e4a3b45bb7c4cacf927a8a, type: 3}
+  m_Name: OnInventoryOpen
+  m_EditorClassIdentifier: 
+  description: 
+  OnEventRaised:
+    m_PersistentCalls:
+      m_Calls: []
