+++ Assets/ScriptableObjects/EventsChannel/Inventory/OnInventoryOpen.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c6bb3ecc4baaf4d4182582a4d7b253f0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
