+++ Assets/Settings/grassToolSettings.asset
@@ -40,10 +40,10 @@ MonoBehaviour:
     m_Bits: 0
   hitMask:
     serializedVersion: 2
-    m_Bits: 2048
+    m_Bits: 0
   paintMask:
     serializedVersion: 2
-    m_Bits: 2048
+    m_Bits: 0
   brushSize: 3.2
   brushFalloffSize: 0.8
   Flow: 0
